consensus: Store transaction nVersion as uint32_t
Summary:
Given that the use of a transaction's nVersion is always as an unsigned
int, it doesn't make sense to store it as signed and then cast it to
unsigned.

This is a partial backport of [[https://github.com/bitcoin/bitcoin/pull/29325 | core#29325]]
https://github.com/bitcoin/bitcoin/pull/29325/changes/27e70f1f5be1f536f2314cd2ea42b4f80d927fbd

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D20608

+++ src/bitcoin-tx.cpp
@@ -244,15 +244,15 @@ static Amount ExtractAndValidateValue(const std::string &strValue) {
 
 static void MutateTxVersion(CMutableTransaction &tx,
                             const std::string &cmdVal) {
-    int64_t newVersion;
-    if (!ParseInt64(cmdVal, &newVersion) ||
+    uint32_t newVersion;
+    if (!ParseUInt32(cmdVal, &newVersion) ||
         newVersion < CTransaction::MIN_VERSION ||
         newVersion > CTransaction::MAX_VERSION) {
         throw std::runtime_error("Invalid TX version requested: '" + cmdVal +
                                  "'");
     }
 
-    tx.nVersion = int(newVersion);
+    tx.nVersion = newVersion;
 }
 
 static void MutateTxLocktime(CMutableTransaction &tx,
